refactor: class WinningNumberDao
LottoGameDao가 가진 WinningLotto 테이블 접근 관련 기능을 옮김
